Added more trade test cases and added basic zombie test cases, along with my solution for each.

+++ src/immortuos/solution/Distance.java
@@ -10,8 +10,8 @@ import immortuos.utils.*;
  */
 public class Distance {
     public static double get(Point p1, Point p2) {
-        double deltaX = (double)(p2.x - p1.x);
-        double deltaY = (double)(p2.y - p1.y);
+        double deltaX = p2.x - p1.x;
+        double deltaY = p2.y - p1.y;
         return Math.sqrt(deltaX * deltaX + deltaY * deltaY);
     }
 }
